adding my algorithm

+++ Entities/Card.cs
@@ -24,8 +24,8 @@ namespace Entities
         [StringLength(int.MaxValue)]
         public string Answer { get; set; }
 
-        public int Repetitions { get; set; }
-        public double EasinessFactor { get; set; }
+        public long Repetitions { get; set; }
+        public decimal EasinessFactor { get; set; }
         public int Interval { get; set; }
 
         public DateTime? PreviousReviewDate { get; set; }
